Move fighter, rogue, and wizard to bandit subdirectory
